\name{rcmap_quick}
\alias{rcmap_quick}
\title{rCrimemap using reformatted JSON data - a faster version of rcmap()}
\usage{
rcmap_quick(period = "2010-12", map_size = c(800, 800),
  map_center = "Lichfield", provider = "Nokia.normalDay", zoom = 7,
  marker = NULL)
}
\description{
## Example Usage:
}
\details{
m_quick <- rcmap_quick() m_quick
}

