% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rcmap_quick}
\alias{rcmap_quick}
\title{rCrimemap using pre-processed JSON data - a faster version of rcmap()}
\usage{
rcmap_quick(period = "2014-01", map_size = c(960, 500),
  map_center = "Lichfield", provider = "OpenStreetMap.BlackAndWhite",
  zoom = 6, marker = NULL)
}
\arguments{
\item{period}{Specific month of interest between Dec 2010 and Jan 2014 in 'yyyy-mm' format (e.g. 2014-01)}

\item{map_size}{Resolution of the map (e.g. Full HD = c(1920 x 1080))}

\item{map_center}{Center of the map, Lichfield is approx. center of England & Wales}

\item{provider:}{Base map service provider (e.g. "OpenStreetMap.BlackAndWhite") (see http://leaflet-extras.github.io/leaflet-providers/preview/index.html)}

\item{zoom}{Zoom level of the map}

\item{marker}{Place a marker at the center? (default = NULL)}
}
\description{
rCrimemap using pre-processed JSON data - a faster version of rcmap()
}
\examples{
## Create a heatmap with default settings
rcmap_quick()

## Be more specific
rcmap_quick(period = "2014-01", map_size = c(1920,1080), provider = "MapQuestOpen.OSM")
}

